Remove di.Parameter, add some tests.
